refactor: simplify architecture by moving config from packages/shared to platform
- Move configuration schema and types from packages/shared/ to platform/src/config/
- Update CLI to import config directly from @startup-gitops/platform
- Remove unnecessary packages/shared/ directory and workspace complexity
- Simplify monorepo to just cli/ and platform/ packages
- Update all imports and TypeScript references
- Reduce dependencies and improve architectural clarity

This change makes the architecture cleaner by having the CLI import configuration
directly from the platform package rather than maintaining a separate shared package.
