+++ cli/package.json
@@ -17,8 +17,7 @@
     "inquirer": "^9.0.0",
     "ora": "^7.0.0",
     "chalk": "^5.0.0",
-    "zod": "^3.0.0",
-    "@startup-gitops/shared": "workspace:*"
+    "@startup-gitops/platform": "workspace:*"
   },
   "devDependencies": {
     "@types/inquirer": "^9.0.0",
