+++ cli/src/commands/destroy.ts
@@ -3,7 +3,7 @@ import { existsSync, readFileSync } from 'fs';
 import inquirer from 'inquirer';
 import ora from 'ora';
 import chalk from 'chalk';
-import { validateConfig, Config } from '@startup-gitops/shared';
+import { validateConfig, Config } from '@startup-gitops/platform';
 import { CDKTFProvider } from '../providers/cdktf';
 import { DigitalOceanProvider } from '../providers/digitalocean';
