+++ cli/tsconfig.json
@@ -8,7 +8,7 @@
   "include": ["src/**/*"],
   "references": [
     {
-      "path": "../packages/shared"
+      "path": "../platform"
     }
   ]
 }
