--- packages/shared/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@startup-gitops/shared",
-  "version": "1.0.0",
-  "description": "Shared types and utilities for GitOps template",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "test": "vitest",
-    "lint": "eslint src --ext .ts"
-  },
-  "dependencies": {
-    "zod": "^3.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.0.0",
-    "typescript": "^5.0.0",
-    "vitest": "^1.0.0",
-    "eslint": "^8.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0"
-  }
-}
