+++ platform/package.json
@@ -18,7 +18,7 @@
     "@cdktf/provider-aws": "^19.0.0",
     "@cdktf/provider-github": "^11.0.0",
     "@cdktf/provider-kubernetes": "^10.0.0",
-    "@startup-gitops/shared": "workspace:*"
+    "zod": "^3.0.0"
   },
   "devDependencies": {
     "@types/node": "^20.0.0",
