+++ platform/src/config/schema.ts
