+++ platform/src/index.ts
