+++ pnpm-workspace.yaml
@@ -1,4 +1,3 @@
 packages:
   - 'cli'
   - 'platform'
-  - 'packages/*'
